+++ fixtures/downstream_project/.formatter.exs
@@ -1,0 +1,4 @@
+[
+  import_deps: [:theoria],
+  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
+]